Unify actions naming - add Action suffix
