+++ prospero-cli/src/main/java/org/wildfly/prospero/cli/ActionFactory.java
@@ -4,35 +4,35 @@ package org.wildfly.prospero.cli;
 import java.nio.file.Path;
 
 import org.jboss.galleon.ProvisioningException;
-import org.wildfly.prospero.actions.ApplyPatch;
+import org.wildfly.prospero.actions.ApplyPatchAction;
 import org.wildfly.prospero.actions.Console;
-import org.wildfly.prospero.actions.InstallationHistory;
-import org.wildfly.prospero.actions.MetadataActions;
-import org.wildfly.prospero.actions.Provision;
-import org.wildfly.prospero.actions.Update;
+import org.wildfly.prospero.actions.InstallationHistoryAction;
+import org.wildfly.prospero.actions.MetadataAction;
+import org.wildfly.prospero.actions.ProvisioningAction;
+import org.wildfly.prospero.actions.UpdateAction;
 import org.wildfly.prospero.api.exceptions.OperationException;
 import org.wildfly.prospero.wfchannel.MavenSessionManager;
 
 public class ActionFactory {
 
-    public Provision install(Path targetPath, MavenSessionManager mavenSessionManager, Console console) {
-        return new Provision(targetPath, mavenSessionManager, console);
+    public ProvisioningAction install(Path targetPath, MavenSessionManager mavenSessionManager, Console console) {
+        return new ProvisioningAction(targetPath, mavenSessionManager, console);
     }
 
-    public Update update(Path targetPath, MavenSessionManager mavenSessionManager, Console console) throws OperationException,
+    public UpdateAction update(Path targetPath, MavenSessionManager mavenSessionManager, Console console) throws OperationException,
             ProvisioningException {
-        return new Update(targetPath, mavenSessionManager, console);
+        return new UpdateAction(targetPath, mavenSessionManager, console);
     }
 
-    public InstallationHistory history(Path targetPath, Console console) {
-        return new InstallationHistory(targetPath, console);
+    public InstallationHistoryAction history(Path targetPath, Console console) {
+        return new InstallationHistoryAction(targetPath, console);
     }
 
-    public MetadataActions metadataActions(Path targetPath) {
-        return new MetadataActions(targetPath);
+    public MetadataAction metadataActions(Path targetPath) {
+        return new MetadataAction(targetPath);
     }
 
-    public ApplyPatch applyPatch(Path targetPath, MavenSessionManager mavenSessionManager, Console console) throws OperationException {
-        return new ApplyPatch(targetPath, mavenSessionManager, console);
+    public ApplyPatchAction applyPatch(Path targetPath, MavenSessionManager mavenSessionManager, Console console) throws OperationException {
+        return new ApplyPatchAction(targetPath, mavenSessionManager, console);
     }
 }
